Add makefile and image folder
The command to compile an asciidoc document with every parameter
correctly set can get a bit overwhelming quickly. To solve this issue we
just use a technology that has been used for decades for exactly these
cases: Makefiles

The Makefile support building pdf and html docs using the build and
image folder as caching storage.

Signed-off-by: Sophie 'Tyalie' Friedrich <dev@flowerpot.me>