Add shell.nix for better portability
Installing and setting up asciidoc with all its dependencies
(especially for asciidoctor-diagram) is rather non-trivial. So to ease
the entry for co-developing this, I provide a shell.nix file.

One can use it by simply typing in nix-shell. The rest is handled in the
background.

To add a new npm package, add it to the node-modules.json and call
node2nix.

Signed-off-by: Sophie 'Tyalie' Friedrich <dev@flowerpot.me>