Introduce numbers and sans-serif for chapters
Changing the font for chapters in contrast to block text, helps add a
nice small visual distinction overall. Also added chapter numbers to
make document it more scrollable.

The document type was also changed to book, so that footnotes are group
by chapters and are not listed all at the end of the document.

Signed-off-by: Sophie 'Tyalie' Friedrich <dev@flowerpot.me>