Introduce docker for building and add mathematical
asciidoctor-mathematical enables rendering of STEM (math stuff) even
for the PDF backend. The default behaviour is to generate MathML code
and let that library handle it online, which is incompatible with PDFs.

asciidoctor-mathematical depends on lasem, which is for one out-dated
and secondly has weird behavior with fonts. After some trial and error
I did not find a solution to make it compatible with shell.nix so the
docker image where it works was used instead.

Signed-off-by: Sophie 'Tyalie' Friedrich <dev@flowerpot.me>