Fix build processes and improve format makefile
When the image directory is the same, changes of the diagram build
pipeline specific to a single backend will not be applied correctly, as
the two backends write over the same images. Especially with cashing
this becomes a problem, so each backend has its own image directory.

Also format makefile for better readibility

Signed-off-by: Sophie 'Tyalie' Friedrich <dev@flowerpot.me>