Add bytefield-svg wrapper for PDF backend
Some things are applicable for all bytefield diagrams like the chosen
font. So that this configuration does not clutter the actual
documentation nor needs endless repeating, I wrote a small wrapper
around bytefield-svg that injects the relevant statements into the
diagram definition. With this it can be handled like a theming choice.

Signed-off-by: Sophie 'Tyalie' Friedrich <dev@flowerpot.me>