Correct color order to BRG
My receiver implementation did st00pid and ignored the first two bytes
of every packet. To still get real looking colors I needed to do a BGR
conversion, but I did not notice the mistake (yes there were image
artifacts, but I did not think much about it).

After a small implementation of the protocol, we found that this did not
work out and that my receiver implementation was wrong.

Signed-off-by: Sophie 'Tyalie' Friedrich <dev@flowerpot.me>
Signed-off-by: Mimoja <mimoja@mimoja.de>