Add descriptions for memory fields
The fields are described in a long yaml file, which at build time is
automatically transpiled into a list of all fields and a large table
which describes the memory. This is done using a custom include
processor written in ruby (in lib/) that converts the yaml.

This complex way of doing things was chosen, because I wanted a space
efficient memory layout for quick referencing, and building this
manually would presumably result in instant inconsistencies.

Signed-off-by: Sophie 'Tyalie' Friedrich <dev@flowerpot.me>